refactor(download kit button): change directory name
change directory name and refactor index.js file
