+++ packages/gatsby-theme-styleguide/src/gatsby-components/designerKitLink/index.js
@@ -1,0 +1,1 @@
+export { default } from './designerKitLink'
